pyproject.toml
[project]
name = "splunk-soar-mcp"
version = "0.1.0"
description = "MCP server for Splunk SOAR On-Prem REST API"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]
